misc/ChekkaBot/challenge/requirements.txt
python-dotenv
discord.py
requests
db-sqlite3